* ivy-emms.el: sort track list and increment version.
At some point, either hash-table-keys or emms-cache-db
changed the order of the keys to the reverse of what it was
before.  This has the effect of presenting album tracks in
reverse order.  Moreover, maybe related, the ordering in
emms-cache-db is by last date added first which,over time,
gets a bit random as new tracks are added.  Thus the
decision to sort the list.

Currently, we sort by alphabetic order of filename (the
hash-key) which work for me.  Raise an issue if it does not
work for you.